Update abstraction alerts to record contact type
We now record the contact type against a notification, starting with abstraction alerts. Because this is the value we'll be persisting into the DB, we stick with all lowercase. If we we need it capitalised we can do so when we display it in the UI. However, when testing the changes we felt no need to.