Reorganise the notify templates object
We need to incorporate 'single use' as a contact type rather than simply recording it under another.

The first step in making that clear is in `app/lib/notify-templates.lib.js`. We update it to follow a constant naming pattern, and ensure the object is frozen to changes, something we should have been doing with our other static lookups.

We then provide the means to identify the template ID to use based on notice type, journey, and contact type. It does mean there is some duplication. But it is much clearer what template is used when.