Update adding additional recipients
Mainly this was to ensure that 'single use' contact type was being set against the additional recipient. But we also now want rhe output to match what we get back from the various Fetch recipients services.

Whilst making the changes, we took the opportunity to make things more consistent between the two services that add additional recipients to the session.