Update recipients service to handle all after fetch
So, prior to this change the `FetchRecipientsService` would call first `RecipientsService` then `DetermineRecipientsService` after each fetch.

`RecipientsService` handled merging the fetched and additional recipients (those added during an ad-hoc notice). `DetermineRecipientsService` then de-duped what `RecipientsService` returned.

First change we made was to rename `RecipientsService` to `MergeRecipientsService` to try and make it clearer what the service was doing.

We also determined this was the best place to copy the return log ID from the fetched recipients, to the additional recipients. They need this information for when they come to be persisted, as these determine which due return logs we look at for assigning due dates.

But we also caught an issue with paper returns download. If you have multiple return logs selected, you'll get the same recipient multiple times in the fetch result.

This is what we want in the download, as we want a line for every return log and recipient combination. However, because `DetermineRecipientsService` was being run _after_ the results had been merged against _all_ of them it was removing the additional return logs from the download data.

It's also not as efficient as it could be. We know the fetched recipients will already have been deduped, i.e. the recipients will be as distinct as we can make them.

So, we only need to ensure that any additional recipients we are adding, don't already exist, either in our fetch recipients or as an existing additional recipient. This means we should only be checking for duplicates when merging the two lists.

Therefore, we've also moved the deduping into `MergeRecipientsService` making `DetermineRecipientsService` redundant.