Implement change to handle differing due dates
Rather than add a lot of complex logic to determine if we have a mix of method types, and then trying to fetch the latest of them we amend the query.

Now, we'll update a return log if the due date is null, or is greater than or equal to the new due date minus one day.