Fix assigning rtn log due dates from notifications (#2747)
https://eaflood.atlassian.net/browse/WATER-5218

We'll now set the due date on return logs when a notification that is linked to the return log is confirmed by [GOV.UK Notify](https://www.notifications.service.gov.uk/) as having been sent.

But we should only be doing this for specific notifications.

Initially, it was succesful returns invitations to licence holders or primary users. However, this wasn't working for ad-hoc returns notices because `return_log_ids` was not being set against any 'single use' contacts added.

We also realised we were not properly handling these 'single-use' contacts. In the absence of knowing what else to set their `message_ref` to, we were using the same value as for licence holders and primary users. This meant the logic could no longer differentiate between 'single use' or a licence holder.

Even worse, we were only checking that the `message_ref` starts with `returns_invitation`. This means notifications to 'returns to' and 'returns agent' contacts were also being picked up for processing.

To resolve these issues, we made the change [Record contact and notice type in notifications](https://github.com/DEFRA/water-abstraction-system/pull/2739).

Now we can see the 'contact type' for a notification, such as 'licence holder', 'single use', or 'returns agent'. We can also tell whether the notice was created via the standard or ad-hoc journeys. We also record the `return_log_ids` against all of them.

Next, we realised we have a problem with paper return notifications. Typically, they are used when a licence is transferred. The new licensee will get the returns invitation. But because the old licensee can no longer access previous return logs online, they need to be sent a paper return.

The problem is that this would not set the 'due date'. It would mean that there would be an ongoing issue of us sending paper returns (essentially an invitation to submit a return), but never recording a due date for them.

So, it was decided that paper returns, as well as return invitations, should set the 'due date' of return logs when confirmed as sent. But now we would want to include notifications to 'single-use' contacts as candidates for processing.

To simplify things, we've moved away from assigning the due date only when the notification is to a licence holder or primary user.

Instead, to cater for both ad-hoc and standard notices, and returns invitations and paper returns, we're trusting 'user knows best'.

Now we will process notifications where

- the notice type (`message_ref`) is paper return or a returns invitation
- the contact type is licence holder, primary user, or single use

Should there be a query about _who_ was sent a notification, we're going to start [displaying the related notifications for a return log](https://github.com/DEFRA/water-abstraction-system/pull/2744).

The final complication is that a user could add both an email and a letter 'single use' contact to an ad-hoc notice. This means we'd be assigning two different due dates to the notifications. The agreement is to use the latter when assigning due dates to the return logs.

This change updates the logic for assigning the due date after checking the notification status to fix the issues highlighted and cover the new requirements.