Fix auto trigger notice not updating status + cnts
https://eaflood.atlassian.net/browse/WATER-5230

As part of our work on setting dynamic due dates for return logs we want to be certain that when we send out invitations to submit return logs that the notification has been received. If the licence is registered, so we have the primary user's email addresses, we will send the returns invitation to it.

If the email fails to send, be it an incorrect email address or a technical issue on our or [GOV.UK Notify's](https://www.notifications.service.gov.uk/) end, we'll then send the returns invitation via letter.

To help the team with this process, we [automated the triggering of sending of the letter](https://github.com/DEFRA/water-abstraction-system/pull/2664.

However, we overlooked calling `UpdateEventService` for the notice after creating it, which means we're not setting its overall status of the counts for it, which means it is displaying incorrectly in the view notices page.

Because this is yet another thing we have to manage when creating and sending an alternate notice, we'll first do some refactoring of `SendNotice` to split out this work. We'll then ensure `UpdateEventService` is called after the notifications have been sent, just like we do for all other notices.