Rename and move update service
Events is the table name but it currently contains more than just notices.

So, we try and use the term 'notice' in preference to make it clear we are talking about the sending of notifications to customers.

We move it because we're about to add a third service that will want to make use of UpdateNoticeService. This means it should be in the root of the notices folder.