Display related communications on a return log (#2744)
https://eaflood.atlassian.net/browse/WATER-5409

As part of our work to set due dates dynamically when we send the notification to the user we have a link to the email/letter to the return log. This enables us to be able to show these notifications on the return log page.

This PR adds in a communication tab to the bottom of the return log view and pulls out some common code for us on both this page and on the view communications page.