Refactor licence purposes to be reusable (#2751)
https://eaflood.atlassian.net/browse/WATER-5381

We are implementing the licence version page, and we will be showing the points / purposes / conditions on the page for that specific licence version. We currently show this per licence.

We have the exising logic in the presenter and view.

This change lifts the logic in the view into a macro so we can use it on the licence versions page. This change also move the presenter logic into the 'LicencePresenter'.