Fix missing due date for paper returns with nulls
https://eaflood.atlassian.net/browse/WATER-5218

In [Fix assigning rtn log due dates from notifications](https://github.com/DEFRA/water-abstraction-system/pull/2747) we updated how due dates are assigned to return logs when notifications are successful. This included setting them when a paper return is successful.

Because of issues with additional contacts in ad-hoc notices, our testing focused on that. So, we were adding aditional contacts to our paper returns and then confirming the due date was being assigned when they were successful.

However, we hadn't spotted that when no additional contacts are added, the `due_date` in the notification is not getting set. We thought it was because we could see it in the preview. However, this is something specific to the preview. We're not setting it when persisting the paper return.

This change ensures we do, which means when we come to assign the due date, we have a value to work with!