Use notification status tag (#2753)
https://eaflood.atlassian.net/browse/WATER-5409

As part of the testing of the new notifications table on the view return log page QA have found that the status tags for the table do not match up with the correct colours as the view notifications page.

Turns out we now have two status tags on the view return log page so this PR updates the view to handle using both.