Ensure due date in failed notice is consistent
When testing we did an ad-hoc returns invitation. For then licence we picked none of its return logs had a null due date, so the one that would have appeared in the notification if it was successfully sent, was the latest.

However, the alternate return invitation calculated the due date as today plus 29 days. This is what would happen if the due date had of been null. But in this case it wasn't.

We realised this means we also needed to extract the due date from the failed notifications, and make a determination as to whether

- we stick with it (as in our test example)
- go with the calculated 'letter' due date (which we would do where due date was null in the return logs)