Update message_ref to use alternate
We thought about using `returns invitation failed` as the message ref in the notification.

However, whilst making these changes we've accepted that `alternate` is a better name for what we are generating. Plus, we already have the column `alternate_notice_id` in the table!

As such, this final change updates the message ref to be `returns invitation alternate`. And as that is effectively our notice type, we also update the notice type to be consistent.