Fix issues with alternate returns invitation (#2759)
https://eaflood.atlassian.net/browse/WATER-5230

As part of our work on setting dynamic due dates for return logs, we want to ensure that when we send invitations to submit return logs, the notifications are received. If the licence is registered, we have the primary user's email address and will send the returns invitation to it.

If the email fails to send, be it an incorrect email address or a technical issue on our or [GOV.UK Notify's](https://www.notifications.service.gov.uk/) end, we'll then send a returns invitation via letter.

To help the team with this process, we [automated the triggering of sending of the letter](https://github.com/DEFRA/water-abstraction-system/pull/2664).

However, in testing, we found some issues with our implementation.

**Not setting overall status**

After sending the notifications linked to a notice, the main process will call `UpdateEventService`. This determines the overall status of the notice and sets the status counts, both of which are used on the view notices page.

When we sent an alternate notice, we were not doing this, so it was not being displayed correctly in the view notices and view notice pages.

**Assigning the wrong alternate notice ID**

When the alternate notice is triggered by another notice with failures, we record the ID of the failed notice against the new alternate notice.

But we also wanted to make it possible to link from a failed notification to the alternate notice, if one was generated. So, we included a query that updates the failed notifications with the new alternate notice ID.

Only we didn't! Instead, we were populating the field with the failed notice ID, essentially duplicating what was already in `event_id`. 🤦

**Not including failed ad-hoc notices**

We were only looking for failed notifications to primary users in standard return invites. Users can now send ad-hoc returns invitations. For example, should licence details change, and the return logs are regenerated, users may wish to resend the invite.

So, we need to ensure we look for failed notifications in both standard and ad-hoc returns invitations.

-----

As part of making these changes, we realised some refactoring was in order. `SendNoticeService` was already doing a lot, so we took this opportunity to split out sending the alternate notice to its own service. This necessitated a 'wrapper' service to manage first sending the 'main' notice, then deciding whether to send the 'alternate' notice — cue lots of renaming and moving of files.

Including ad-hoc returns invitations also forced us to consider the due date we use. In standard returns invitations, all due dates are calculated. Emails it is today + 28, so when we send the alternate notice, we want to use today + 29 (these are the agreed-upon due dates based on whether it is an email or a letter).

However, in ad-hoc, if all the due return logs for the selected licence _have_ a due date, we are required to use the _latest_ in the notification. Therefore, the same should apply to the alternate notice we generate when the ad-hoc fails. So, now `FetchFailedReturnsInvitationsService` returns the due date to assign to the new notifications.

Other than some general housekeeping, the final change to call out is ensuring we are consistent with our notice type and message ref. We were mixing the terms 'failed' and 'alternate'. Carrying out this work, we decided 'alternate' was the most appropriate name, not least because we named the field we update in failed notifications `alternate_notice_id`!