Remove wait in send alternate notice  (#2762)
https://eaflood.atlassian.net/browse/WATER-5230

As part of our work on setting dynamic due dates for return logs, we want to ensure that when we send invitations to submit return logs, the notifications are received. If the licence is registered, we have the primary user's email address and will send the returns invitation to it.

If the email fails to send, be it an incorrect email address or a technical issue on our or [GOV.UK Notify's](https://www.notifications.service.gov.uk/) end, we'll then send a returns invitation via letter.

To help the team with this process, we [automated the triggering of sending of the letter](https://github.com/DEFRA/water-abstraction-system/pull/2664).

However, in testing, we found some issues with our implementation.

These were [fixed](https://github.com/DEFRA/water-abstraction-system/pull/2759) but we've realised that

- We are pausing after sending the alternate notice unnecessarily
- We are calling `UpdateNoticeService` twice when we could call it once after both the main and alternate notice have been created

This is a housekeeping change to resolve these issues.