Remove feature flag ENABLE_NULL_DUE_DATE (#2763)
https://eaflood.atlassian.net/browse/WATER-4991
https://github.com/DEFRA/water-abstraction-team/issues/140

This feature is about to go live and needs to be up and running by the time the quarterly returns notice is issued in the new year.

We have to apply changes to the existing records as well, so 'flipping back' is not really an option. So, on the eve of its release, we remove the feature flag to reduce maintenance and the risk of unexpected issues.