Update conditions to use licence version (#2765)
* Update conditions to use licence version

https://eaflood.atlassian.net/browse/WATER-5381

The licence conditions need to use the current licence version. This has been updated previously using a modifier.

The current conditions query was basing its licence version of the status. This is wrong and needs to be updated.

This change updates the licence conditions fetch to use the current licence version.