Update licence version to show conditions (#2768)
https://eaflood.atlassian.net/browse/WATER-5381

The licence version page needs to show the conditions for the licence version.

This can result in not showing conditions for a licence version, when this is the case we show our standard 'No conditions for x' message.

This change adds the conditions to the licence version page, and uses the existing conditions  presenter and macro to render the conditions.