Update fetch conditions (remove 'current') (#2772)
https://eaflood.atlassian.net/browse/WATER-5381

We updated the fetch conditions query to use the licence version id and not the licence.

This allows the conditions pages service to pass in the 'current' licence version, and allows the licence version page to pass in the licence version id selected.

This works because the conditions page in the 'currentLicenceVersion'. This meant we needed could remove 'andWhere('licence_version.status', 'current')' form the query.

However, when we made this change we missed an additional 'andWhere('licence_version.status', 'current')' when get the licence version purpose. This resulted in the condition type being returned with no conditions.

This change removes the '.andWhere('licence_version.status', 'current')' from the licence version purpose.

An additional test has been to check not conditions are returned.