Billing account change address - Select billing account (#2770)
https://eaflood.atlassian.net/browse/WATER-5396

As part of our ongoing work to decommission our legacy UI we identify journeys that are small enough to be ported over in discreet chunks of work. This is part one of the change billing address journey and is mostly for setting up the scaffolding of the controller, routes, folder structure and initiating of the session.

The more complex logic of fetching the data and displaying the company information will come in a subsequent PR.