Update licence page to show issue date
https://eaflood.atlassian.net/browse/WATER-5420

The licence summary page need to show the issue date, work has been done recently to bring this into the database.

This change adds the 'issueDate' to the licence summary page. This follows the same pattern as the 'startDate' where we need to check if the there is current licence version.