Fix saving of profile details
The profile details page doesn't save the details when the user doesn't already have a `contactDetails` object in the `userData` attribute, or if the `userData` attribute is `null`.

This change fixes that, by patching the existing value to add the necessary containing objects before setting attributes of the `contactDetails` object.