Billing account change address - Fetch company details (#2784)
https://eaflood.atlassian.net/browse/WATER-5421

As part of our ongoing work to decommission our legacy UI we identify journeys that are small enough to be ported over in discreet chunks of work.

This PR fetches the billing account information for the existing billing account and adds it to the session as it will be used through out the journey. It also adds a feature flag for the link to this journey.