Select an existing address scaffold (#2796)
https://eaflood.atlassian.net/browse/WATER-5432

As part of our ongoing work to decommission our legacy UI we identify journeys that are small enough to be ported over in discreet chunks of work. This is part of the change billing address journey and is mostly for setting up the scaffolding of the controller, routes and services of the selecting an existing address page.

The more complex logic of fetching the data and displaying the addresses will come in a subsequent PR.