Fix return log ID name confusion https://eaflood.atlassian.net/browse/WATER-5427
For reasons we’ll never know, the previous team opted to generate the IDs for return logs based on data within the record, for example, `v1:2:03/28/77/0028:10046670:2023-04-01:2024-02-05`.

For those in the know, they’ll see the NALD region, licence reference, return requirement number, start and end date for the return log. This is fine as a reference, but a really bad idea as the unique ID for the record.

When we took over returns management, we added a proper ID to the table, and since it was implemented it is what we have been using in the service when working with return logs.

In a previous PR https://github.com/DEFRA/water-abstraction-system/pull/2791 the return submission has been joined to the return log via the new `return_id`.

All that remains to do in this PR is to remove the aliasing from the return log IDs in our views so that our new ID is called `id` rather than `return_id`. Then fix all the code that this change breaks.