Billing account change address - Select account  - add search input (#2798)
https://eaflood.atlassian.net/browse/WATER-5421

As part of our ongoing work to decommission our legacy UI we identify journeys that are small enough to be ported over in discreet chunks of work.

This PR adds a search input to the select account page and saves it to the session. This will be used in a subsequent page to find a different account to use.