Billing account change address - select an existing address (#2803)
https://eaflood.atlassian.net/browse/WATER-5432

As part of our ongoing work to decommission our legacy UI we identify journeys that are small enough to be ported over in discreet chunks of work. This is part of the change billing address journey.

This PR fetches the associated address from the company id and displays them to the user. The user can then choose one and it will save it to the session.