Update licence versions pagination order by to use 'endDate' (#2807)
https://eaflood.atlassian.net/browse/WATER-5401

When we implemented the original work for paginating the licence version page we missed an edge case where the start date is in the same year.

When the licence versions where in this state the end date was not accommodated, this resulted with some of the licence version appealing in the wrong order.

This change updates the fetch licence versions query to order by the start date, then the end date.