Link return submissions to logs via UUID (#2791)
https://eaflood.atlassian.net/browse/WATER-5425

For reasons we’ll never know, the previous team opted to generate the IDs for return logs based on data within the record, for example, `v1:2:03/28/77/0028:10046670:2023-04-01:2024-02-05`.

For those in the know, they’ll see the NALD region, licence reference, return requirement number, start and end date for the return log. This is fine as a reference, but a really bad idea as the unique ID for the record.

When we took over returns management, we added a proper ID to the table, and since it was implemented it is what we have been using in the service when working with return logs.

However, return submissions are still linked to return logs by the wonky return_id.

The task for this PR is to:
- Add the new column `return_log_id` (UUID) to `return_submissions` view and legacy `returns.versions` table via a migration.
- Wherever a return submission is created, ensure `return_log_id` is populated with the `id` of the `return log`, as well as continuing to populate `return_id`.
- Update the model relationships between `ReturnLogModel` and `ReturnSubmissionModel` to use the new field.
- Update any custom and ad-hoc fetches that rely on linking return submissions to return logs via the old field to use the new one.