Add pagination to customers licences page
https://eaflood.atlassian.net/browse/WATER-5384

We recently implemented the customers licences page. This did not include pagination as we were working on a updating the pagination presenter to return the standard pagination message of 'Showing X of Y'

This change uses this updated pagination presenter to add pagination to the page. The fetch has been updated to provide the pagination object.