Add latest and due date status to generate query
This change adds the two new values we need to the query. They are done via new CTEs and the results then added to the final output.

At the same time we also did some tidying up. Some things were not in alphabetical order, the write case, or were being aliased unnecessarily.

The final thing we fixed was the output for the download. We spotted that if return logs shared the same licence ref, recipient and return reference, only one was being returned, even though the rest of the details were different.

So, making them distinct only on contact_hash_id and return_id ensures the distinct is doing what we wanted, removing dupes in the recipients whilst ensuring every relevant return log is included in the results.