Remove latest due date handling from SubmitLicence
Prior to this change, we'd work out the latest due date from the return logs fetched when the licence has been selected in the ad-hoc journey.

We'd save this in the session for use later by the NotificationsPresenter, specifically for handling ad-hoc returns invitations.

Now the Fetch notice services are determining what the notification due date should be, using the latest due date returned by the fetch query, there is no need for use to be performing this work any more.