Update download returns notice presenter
Update it to include the new fields.

We started fetching return ID when we spotted that some return logs returned by the queries generated were not being included in the results.

However, we realised it was handy to have in the results as it provides an easy means to find the matching return log quickly in the UI (just update the ID in the URL).

notificationDueDate was added because we cannot currently see the determined due date in the template. We removed it temporarily whilst this change is being developed just in case a user attempts to generate a returns reminder before we're done.

But again, we felt it would be useful for users to have as a record of what due date was sent in the notifications to the recipients.