Housekeeping - Rename notifications presenter
With only `NotificationsPresenter` as a name, you'd be forgiven for thinking that this presenter is used when creating _all_ notifications. In fact, it is just used when creating returns invitation and reminder notifications, which we have grouped elsewhere as `returns-notice` notifications.

So, we rename the presenter to make this clearer.