Fix - Handle unable to connect
To test the alternate returns invitation (when the email to the primary user fails) we generally replace the API key with a bogus one.

But we also tried replacing the API URL, and in that case exposed an issue with the error handling. The presenter is assuming a connection is always made, and therefore Notify will return a body.

But if you fail to connect, then the response object won't contain a body. This causes the presenter to throw an error which means the notification never gets marked as 'errored'.