Update FetchPaperReturnsRecipientsService
To be consistent with the new FetchReturnsInvitationRecipients and FetchReturnsReminderRecipients services, and to ensure we support new fields we'll be adding to the download, we need to update FetchPaperReturnsRecipientsService to also ensure `notificationDueDate` is set
by the service.

It is a little different in that the presenter still needs to determine the due date when it comes to _sending_ the notification, because when sending it is not until later that the recipients and return logs are combined.

But when fetching the data for the download, the results are per recipient and return log combination, so we can set the value.