Filter out registered licences in json_contacts
Another bit of de-duping that also should improve performance slightly.

Instead of filtering out registered licences in both the licence holder and returns to parts of the query, we do it once in the json_contacts CTE.

It also avoids us extracting the JSONB contacts, and generating hashes for contacts we would previously have then filtered out.