Always use latest date for return reminder notices (#2812)
https://eaflood.atlassian.net/browse/WATER-5305
https://eaflood.atlassian.net/browse/WATER-5357

The original plan for return reminder notices was to stop displaying the 'due date' on the reminder.

Because 'recipient' is calculated when the notice is sent, it's not possible to know which notifications a recipient has previously received. Therefore, what 'due date' was on the invitation?

However, it has now been decided that displaying a 'due date' in the notification is the key imperative to action, so it needs to be brought back.

This means the query for returns reminders now needs to select the latest 'due date' for each recipient, from the return logs associated with them.

This means the logic for returns invitations and reminders is different. It has led to a refactoring of how and when the 'due date' is determined.

We've replaced fetching by journey with fetching recipients by notice type. This means we can move the logic to assign the due date to the fetch services.

It also means we are assigning the 'notification due date' when fetching the recipient, instead of when generating the notification via the presenter. That also meant we needed to know the latest due date in the query to support the ad-hoc returns invitation journey. That meant we no longer need to determine it when the licence number is selected at the start of the setup journey.

We added 'notification due date' to the download to help us with testing, but have decided to keep it as it will be helpful for users' reference. We also noticed that some return logs were not included in the download because the distinct was not working correctly. Using `return_id` not only fixes the problem, but also means for those that know how our URL's work, they can go straight to a return log listed in the download.