Keep the authenticated session alive in the new application (#2813)
https://eaflood.atlassian.net/browse/WATER-5438

Currently, the cookie used to track a user's authenticated session doesn't have its TTL refreshed in the new `water-abstraction-system` app. This causes the user's session to timeout if they don't visit a page from the old `water-abstraction-ui` application.

This change updates the authentication configuration to keep the session alive, updating the TTL of the cookie.

Prior to this change, we were only _reading_ the contents of the cookie, but we are now _writing_ its value, so we have to explicitly set its path to the root `/` of the application that is proxying to this application, as otherwise it will only be set for the `/system` path that this application uses as its base path.