Housekeeping - Fix up currentVersion() tests
The tests were creating multiple licence records unnecessarily, once in the top level beforeEach() and then in each subsequent beforeEach().

It also wasn't using the naming pattern used for `$licenceHolder()`. That creates an object called `licence`, uses it's ID in the setup, and then sets `testRecord` to the result of fetching the matching record along with applying the associated modifier.

It was also creating versions that were never used in the top beforeEach().