Housekeeping - Further fixups for currentVersion()
We've updated the tests to better reflect real-life data. For example, superseded versions would also ways be inserted before 'current' version, this better reflects the order in tables should a query be unordered.

Also, the dates, issue and increment always align. Rather than leave that to the helper, we ensure they are set to reflect how the records would actually be populated.