Fix licence $currentVersion() modifier (#2814)
https://eaflood.atlassian.net/browse/WATER-5428

The business reported that the purpose shown on the licence summary page for some licences was incorrect.

What was shown was the purpose of a replaced licence version.

The problem was caused by a ‘fix’ we made to the summary screen.

Previously, we showed the data for the licence version flagged as `current` on the summary page. This is how the legacy page did it, so we copied it when rebuilding it.

As part of 'view my licence 2.0', it was spotted that in some cases a new licence version is created that starts in the future. So, though it has the status of `current`, it is not the 'current' licence version.

We tweaked our logic to ensure the version we show is the 'current' one for that time. But the users had found a scenario that was not being handled correctly.

For both reported licences, their latest licence versions start on the same date: 2022-04-01. For example, imagine these are the latest licence version records for a licence.

|Licence Ref|Version Id|Start date|End date  |Issue|Increment|Status |
|-----------|----------|----------|----------|-----|---------|-------|
|01/123     | 007      |2022-04-01|          | 5   | 2       |current|
|01/123     | 006      |2022-04-01|2022-04-01| 5   | 1       |current|

The old logic said `SELECT * WHERE status = ‘current’;` Now we are doing `SELECT * WHERE start_date <= TODAY ORDER BY start_date DESC LIMIT 1;`.

So, we’re taking the first result where the start date is less than the current date. For these licences, it means you could get `006` returned, when obviously you want `007`.

In fact, if you refreshed the page, you could see the purpose change to the correct one, and when refreshed, go back to the old one! It’s not guaranteed, but that is part of the problem; we should control the order of the results.

It’s a quick fix. If we also order by issue and increment, we ensure we get the latest version for a given date, and, in the case of these licences, the correct licence version.